+++ packages/appcd-gulp/CHANGELOG.md
@@ -1,3 +1,10 @@
+# v3.1.5
+
+ * fix(watch): Improved error handling during `watch` task.
+ * fix(lint): Added appcd's `node_modules` to lookup paths when `appcd-gulp` is symlinked (via
+   `yarn link` in this case) into another project.
+ * style(test): Updated code coverage HTML report to use a dark mode color palette.
+
 # v3.1.4 (Mar 3, 2021)
 
  * misc: Improved `package` task support for `__dirname` and `__filename`, but still doesn't